refactor tests of trailing slash behaviour into separate module
