+++ tests/test_attributes.py
@@ -7,16 +7,6 @@ from app.api import crud
 from app.api import utility as util
 
 
-def test_root(test_app):
-    """Given a GET request to the root endpoint, Check for 200 status and expected content."""
-
-    response = test_app.get("/")
-
-    assert response.status_code == 200
-    assert "Welcome to the Neurobagel REST API!" in response.text
-    assert '<a href="/docs">documentation</a>' in response.text
-
-
 @pytest.mark.filterwarnings("ignore:.*NB_API_ALLOWED_ORIGINS")
 @pytest.mark.parametrize(
     "valid_data_element_URI",
@@ -242,12 +232,3 @@ def test_get_attribute_vocab(
         "namespace_prefix": expected_namespace_pfx,
         "term_labels": mock_term_labels,
     }
-
-
-def test_request_with_trailing_slash_not_redirected(test_app):
-    """
-    Test that a request to a route with a trailing slash '/', where none is expected,
-    is *not* redirected and returns a 404.
-    """
-    response = test_app.get("/attributes/nb:SomeClass/")
-    assert response.status_code == 404
