+++ tests/test_query.py
@@ -511,21 +511,3 @@ def test_query_without_token_succeeds_when_auth_disabled(
     monkeypatch.setattr(crud, "get", mock_successful_get)
     response = test_app.get(ROUTE)
     assert response.status_code == 200
-
-
-def test_request_without_trailing_slash_not_redirected(
-    test_app, monkeypatch, mock_successful_get, disable_auth
-):
-    """Test that a request to the /query route is not redirected to have a trailing slash."""
-    monkeypatch.setattr(crud, "get", mock_successful_get)
-    response = test_app.get(ROUTE, follow_redirects=False)
-    assert response.status_code == 200
-
-
-def test_query_params_with_leading_slash_fails(test_app, disable_auth):
-    """
-    Test that a request to the /query route with a slash before query parameters
-    is *not* redirected to exclude the slash, and fails with a 404.
-    """
-    response = test_app.get(f"{ROUTE}/?min_age=20")
-    assert response.status_code == 404
